requirements.txt
pycti
python-dotenv
pylint
